22172 - Mcollective ssh key security should allow a custom per-identity authorized keys file
Node verifier should never use ssh agent.

mcollective is a long running process which is run as a daemon. It's never correct for the _server_
to use the user who starts it's agent forwarding to verify things with, as this will stop working
as soon as the user who started the daemon logs out